requirements.txt
openai>=1.60.0
google-genai>=1.0.0
streamlit>=1.37.0
Pillow>=10.0.0
